+++ packages/core/src/store/api/baseApi.ts
@@ -41,6 +41,7 @@ export const baseApi = createApi({
     'Entity',
     'Space',
     'SpaceMember',
+    'TableRow',
     // Future tag types:
     // 'Comment',
   ],